+++ packages/base/dataLoader.js
@@ -160,10 +160,10 @@ function loadArrayData (conn, dataArray, entityName, ettAttributes, mapping, tra
     entity: 'ubm_enum',
     keyAttribute: 'eGroup;code',
     localization: [
-      {keyValue: 'UBS_MESSAGE_TYPE;user',  execParams: {name: 'Користувачів'}},
-      {keyValue: 'UBS_MESSAGE_TYPE;system',  execParams: {name: 'Система'}},
-      {keyValue: 'UBS_MESSAGE_TYPE;warning',  execParams: {name: 'Попереждення'}},
-      {keyValue: 'UBS_MESSAGE_TYPE;information',  execParams: {name: 'Інформація'}}
+      {keyValue: 'UBS_MESSAGE_TYPE;user',  execParams: {name: 'Users'}},
+      {keyValue: 'UBS_MESSAGE_TYPE;system',  execParams: {name: 'System'}},
+      {keyValue: 'UBS_MESSAGE_TYPE;warning',  execParams: {name: 'Warning'}},
+      {keyValue: 'UBS_MESSAGE_TYPE;information',  execParams: {name: 'Info'}}
     ]
  }
  loader.localizeEntity(session, localizationConfig, __filename)