+++ packages/stubs/_signatureValidationResult-stub.js
@@ -10,7 +10,7 @@
  * @property {boolean} ocspVerified Is signer certificate status verified during signing
  * @property {boolean} hardwareKeyUsed Is hardware token used to made signature
  * @property {Date} signingTime Time of signing. In case tspValid this time is taken from authorised source using TSP protocol, in other case - this is **UNTRUSTED** time of local signer computer
- * @property {Boolean} isDigitalStamp Is signature certificate is a digital time stamp (печать)
+ * @property {Boolean} isDigitalStamp Is signature certificate is a digital time stamp
  *
  * @property {Object} certificate Signer certificate information. In case valid===false certificate may not exists (or broken) and this property is empty object
  * @property {String} certificate.keyUsage String with key usage information