+++ packages/ubs/public/index.js
@@ -1,5 +1,4 @@
 require('./dataBinder')
-require('./cliUtils')
 require('./css/ubs.css')
 require('./settings')
 require('./UBReport')