+++ packages/adminui-vue/components/controls/UCron/dayCron.vue
@@ -1,0 +1,18 @@
+<template>
+  <seconds-cron
+    v-bind="$attrs"
+    mode="day"
+    :length="30"
+    :start-count="1"
+    v-on="$listeners"
+  />
+</template>
+
+<script>
+export default {
+  name: 'DayCron',
+  components: {
+    SecondsCron: require('./secondsCron.vue').default
+  }
+}
+</script>